🧹 Tidy: 共通定数の整理とFatなuseDataフックの分割
- **定数の共通化**: 授乳履歴とオムツ履歴で使用されるラベルやスタイル定義をコンポーネント内から `frontend/constants/` 配下の専用ファイル (`feeding.ts`, `diaper.ts`) に移動し、再利用可能にしました。
- **フックの分割**: 肥大化していた `useData.ts` から、ドメインごとに責務を分割した新しいフック (`useBabies`, `useRecords`, `useFamily`) を切り出しました。
- **後方互換性**: `useData.ts` は新しいフックを再エクスポートする形に変更し、既存のインポートを壊さないように配慮しました。

これにより、コードの重複が排除され、各ファイルが単一の責務を持つようになり、将来的な保守性が向上します。
