+++ frontend/hooks/useBabies.ts
@@ -1,0 +1,19 @@
+import useSWR from 'swr';
+import { fetcher } from '@/lib/api';
+import type { Baby } from '@/types/baby';
+
+export function useBabies(options?: { fallbackData?: Baby[] }) {
+    const { data, error, isLoading, mutate } = useSWR<Baby[]>('/babies/', fetcher, {
+        ...options,
+        shouldRetryOnError: false,
+        revalidateOnFocus: true,
+        revalidateOnMount: true,
+        keepPreviousData: true,
+    });
+    return {
+        babies: data,
+        isLoading,
+        isError: error,
+        mutate,
+    };
+}
